Refactor test and Added msbuild flags (#2113)
