+++ conans/test/build_helpers/msbuild_test.py
@@ -1,0 +1,59 @@
+import platform
+import unittest
+
+from nose.plugins.attrib import attr
+
+from conans.paths import CONANFILE
+from conans.test.utils.tools import TestClient
+from conans.test.utils.visual_project_files import get_vs_project_files
+
+
+class MSBuildTest(unittest.TestCase):
+
+    @attr('slow')
+    def build_vs_project_test(self):
+        if platform.system() != "Windows":
+            return
+        conan_build_vs = """
+from conans import ConanFile, tools, MSBuild
+import platform
+
+class HelloConan(ConanFile):
+    name = "Hello"
+    version = "1.2.1"
+    exports = "*"
+    settings = "os", "build_type", "arch", "compiler"
+
+    def build(self):
+        msbuild = MSBuild(self)
+        msbuild.build("MyProject.sln")
+
+    def package(self):
+        self.copy(pattern="*.exe")
+
+"""
+        client = TestClient()
+
+        files = get_vs_project_files()
+        files[CONANFILE] = conan_build_vs
+
+        # Try with x86_64
+        client.save(files)
+        client.run("export lasote/stable")
+        client.run("install Hello/1.2.1@lasote/stable --build -s arch=x86_64")
+        self.assertIn("Release|x64", client.user_io.out)
+        self.assertIn("Copied 1 '.exe' files: MyProject.exe", client.user_io.out)
+
+        # Try with x86
+        client.save(files, clean_first=True)
+        client.run("export lasote/stable")
+        client.run("install Hello/1.2.1@lasote/stable --build -s arch=x86")
+        self.assertIn("Release|x86", client.user_io.out)
+        self.assertIn("Copied 1 '.exe' files: MyProject.exe", client.user_io.out)
+
+        # Try with x86 debug
+        client.save(files, clean_first=True)
+        client.run("export lasote/stable")
+        client.run("install Hello/1.2.1@lasote/stable --build -s arch=x86 -s build_type=Debug")
+        self.assertIn("Debug|x86", client.user_io.out)
+        self.assertIn("Copied 1 '.exe' files: MyProject.exe", client.user_io.out)
