+++ conans/test/util/tools_test.py
@@ -5,7 +5,6 @@ import unittest
 from collections import namedtuple
 
 from conans.client.client_cache import CONAN_CONF
-from nose.plugins.attrib import attr
 
 from conans import tools
 from conans.client.conan_api import ConanAPIV1
@@ -13,11 +12,11 @@ from conans.client.conf import default_settings_yml, default_client_conf
 
 from conans.errors import ConanException
 from conans.model.settings import Settings
-from conans.paths import CONANFILE
+
 from conans.test.utils.runner import TestRunner
 from conans.test.utils.test_files import temp_folder
 from conans.test.utils.tools import TestClient, TestBufferConanOutput
-from conans.test.utils.visual_project_files import get_vs_project_files
+
 from conans.test.utils.context_manager import which
 from conans.tools import OSInfo, SystemPackageTool, replace_in_file, AptTool, ChocolateyTool,\
     set_global_instances
@@ -413,54 +412,6 @@ compiler:
         with tools.environment_append({"CONAN_BASH_PATH": "path\\to\\mybash.exe"}):
             tools.run_in_windows_bash(conanfile, "a_command.bat")
             self.assertIn("path\\to\\mybash.exe --login -c", conanfile.command)
-
-    @attr('slow')
-    def build_vs_project_test(self):
-        if platform.system() != "Windows":
-            return
-        conan_build_vs = """
-from conans import ConanFile, tools, MSBuild
-import platform
-
-class HelloConan(ConanFile):
-    name = "Hello"
-    version = "1.2.1"
-    exports = "*"
-    settings = "os", "build_type", "arch", "compiler"
-
-    def build(self):
-        msbuild = MSBuild(self)
-        msbuild.build("MyProject.sln")
-
-    def package(self):
-        self.copy(pattern="*.exe")
-
-"""
-        client = TestClient()
-
-        files = get_vs_project_files()
-        files[CONANFILE] = conan_build_vs
-
-        # Try with x86_64
-        client.save(files)
-        client.run("export lasote/stable")
-        client.run("install Hello/1.2.1@lasote/stable --build -s arch=x86_64")
-        self.assertIn("Release|x64", client.user_io.out)
-        self.assertIn("Copied 1 '.exe' files: MyProject.exe", client.user_io.out)
-
-        # Try with x86
-        client.save(files, clean_first=True)
-        client.run("export lasote/stable")
-        client.run("install Hello/1.2.1@lasote/stable --build -s arch=x86")
-        self.assertIn("Release|x86", client.user_io.out)
-        self.assertIn("Copied 1 '.exe' files: MyProject.exe", client.user_io.out)
-
-        # Try with x86 debug
-        client.save(files, clean_first=True)
-        client.run("export lasote/stable")
-        client.run("install Hello/1.2.1@lasote/stable --build -s arch=x86 -s build_type=Debug")
-        self.assertIn("Debug|x86", client.user_io.out)
-        self.assertIn("Copied 1 '.exe' files: MyProject.exe", client.user_io.out)
 
     def download_retries_test(self):
         out = TestBufferConanOutput()
